nfa-state: reduce storage for state node marker
The state node marker should be capable to mark split node, plus upto 256
character nodes. Reduce storage for to allow extra node fields with same
node storage size, and to fit node to CPU cache line on common CPUs.

+++ nfa-state.h
@@ -17,7 +17,7 @@ enum nfa_type {
 
 struct nfa_state {
 	struct nfa_state *next;
-	int c;
+	short c;
 	int index;	/* used by processor to map state to state index */
 	struct nfa_state *out[2];
 };
